alysida-platform/requirements.txt
falcon
gunicorn
httpie
pycrypto